# backend/openapi.json
{
  "components": {
    "schemas": {
      "Actor": {
        "description": "IF-02, Decision R2-7. Who is behind a write performed under a role.\n\nAbsent means `agent`; the orchestrator always sends `agent` and never claims to be a\nhuman, which is what makes the human gate on collisions (FR-OPS-07) mean anything.",
        "enum": [
          "agent",
          "human"
        ],
        "title": "Actor",
        "type": "string"
      },
      "AgentRole": {
        "description": "FR-PERM-01, transcribed from Figure 3 and in its order.",
        "enum": [
          "architect",
          "world_builder",
          "writer",
          "style_editor",
          "auditor",
          "canoniser"
        ],
        "title": "AgentRole",
        "type": "string"
      },
...
      "ChangesFile": {
        "additionalProperties": false,
        "description": "DR-08, DR-10. The whole of `cast/{id}/changes.yaml`.\n\nA list, because a character can lose the same attribute twice - a hand, then the\nprosthesis - and each step has its own scene and its own cause.",
        "properties": {
          "changes": {
            "description": "Every registered change for this character; the auditor reads those dated at or before the scene under audit.",
            "items": {
              "$ref": "#/components/schemas/ChangeEvent"
            },
            "title": "Changes",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "title": "ChangesFile",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProvenanceRecord": {
        "description": "One store write. Every field answers a question someone asks after the fact.",
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/Actor",
            "description": "`agent` or `human` (Decision R2-7). A human acts under a role rather than beside it, which is why there is no `human` row in Figure 3."
          },
          "at": {
            "description": "ISO-8601 UTC timestamp.",
            "title": "At",
            "type": "string"
          },
          "content_hash": {
            "description": "SHA-256 of the bytes written, hex.",
            "title": "Content Hash",
            "type": "string"
          },
          "path": {
            "description": "Store-relative POSIX path that was written.",
            "title": "Path",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/AgentRole",
            "description": "The Figure 3 role the write was performed under."
          },
          "scene": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Scene id, when the write is in a turn.",
            "title": "Scene"
          },
          "turn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Turn id, when the write is in a turn.",
            "title": "Turn"
          }
        },
        "required": [
          "path",
          "role",
          "actor",
          "content_hash",
          "at"
        ],
        "title": "ProvenanceRecord",
        "type": "object"
      },
...
      "RelationshipsFile": {
        "additionalProperties": false,
        "description": "DR-08, DR-10. The whole of `cast/relationships.yaml`.\n\nOne file for every edge in the cast. It is not split per character because each edge\nwould then have two homes and one of them would fall behind.",
        "properties": {
          "relationships": {
            "description": "Every directed edge in the cast; both directions of a pair are separate rows and may disagree.",
            "items": {
              "$ref": "#/components/schemas/Relationship"
            },
            "title": "Relationships",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "title": "RelationshipsFile",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "The only process that reads and writes the harness stores.",
    "title": "My Story Marker \u2014 backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cast/relationships": {
      "get": {
        "description": "IF-03, `GET /cast/relationships`. One file for the whole graph: an edge is directed,\nand the two directions of a pair are separate rows that may disagree.",
        "operationId": "read_relationships_cast_relationships_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationshipsFile"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read every relationship in the cast",
        "tags": [
          "cast"
        ]
      },
      "put": {
        "description": "IF-04, `PUT /cast/relationships` (canoniser).\n\nThe whole file at once, because an edge belongs to neither end of itself and there is no\nper-character slice of it to write.",
        "operationId": "write_relationships_cast_relationships_put",
        "parameters": [
          {
            "description": "The Figure 3 role performing this write. Required on every write.",
            "in": "header",
            "name": "X-Agent-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The Figure 3 role performing this write. Required on every write.",
              "title": "X-Agent-Role"
            }
          },
          {
            "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipsFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvenanceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write every relationship in the cast",
        "tags": [
          "cast"
        ]
      }
    },
    "/cast/{id}": {
      "get": {
        "description": "IF-03, `GET /cast/{id}`. The complete record.\n\n`GET /cast/{id}/dossier?at=` below is the same record trimmed to an instant (FR-OPS-01),\nand this one is its unsafe sibling: a writer handed the complete dossier uses facts the\ncharacter has not yet learned, because nothing in the text marks them as future. The\nrecord reads as true, and everything true in the context is fair to write. This route is\nfor a human reading their own tree, never for an assembled context.",
        "operationId": "read_character_cast__id__get",
        "parameters": [
          {
            "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "maxLength": 120,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_-]*$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a character's dossier",
        "tags": [
          "cast"
        ]
      }
    },
    "/cast/{id}/changes": {
      "get": {
        "description": "IF-03, `GET /cast/{id}/changes`. The register that tells a scar in chapter nine apart\nfrom a continuity error (invariant 3).",
        "operationId": "read_changes_cast__id__changes_get",
        "parameters": [
          {
            "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "maxLength": 120,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_-]*$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangesFile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a character's registered changes",
        "tags": [
          "cast"
        ]
      },
      "put": {
        "description": "IF-04, `PUT /cast/{id}/changes` (canoniser).",
        "operationId": "write_changes_cast__id__changes_put",
        "parameters": [
          {
            "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "maxLength": 120,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_-]*$",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "description": "The Figure 3 role performing this write. Required on every write.",
            "in": "header",
            "name": "X-Agent-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The Figure 3 role performing this write. Required on every write.",
              "title": "X-Agent-Role"
            }
          },
          {
            "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangesFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvenanceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write a character's registered changes",
        "tags": [
          "cast"
        ]
      }
    },
    "/cast/{id}/dossier": {
      "get": {
        "description": "FR-OPS-01, IF-03 `GET /cast/{id}/dossier?at=`. The character as they were at `at`.\n\nOnly the facts already acquired, the valence of each relationship on that date, the point\non the arc, and the body with every registered change up to then applied (AC 10). The\nload-bearing call of assembly: withholding a future fact is more reliable than\ninstructing the model to ignore it.",
        "operationId": "read_dossier_at_cast__id__dossier_get",
        "parameters": [
          {
            "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "maxLength": 120,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_-]*$",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "description": "The story instant to trim to: integer hours since `epoch_zero` (Decision 8), negative for prequel scenes. Required, because there is no safe default: 'now' on the story axis is whatever the scene being written says it is.",
            "in": "query",
            "name": "at",
            "required": true,
            "schema": {
              "description": "The story instant to trim to: integer hours since `epoch_zero` (Decision 8), negative for prequel scenes. Required, because there is no safe default: 'now' on the story axis is whatever the scene being written says it is.",
              "title": "At",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrimmedDossier"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a character as of a story instant",
        "tags": [
          "cast"
        ]
      },
      "put": {
        "description": "IF-04, `PUT /cast/{id}/dossier` (canoniser).",
        "operationId": "write_dossier_cast__id__dossier_put",
        "parameters": [
          {
            "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "maxLength": 120,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_-]*$",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "description": "The Figure 3 role performing this write. Required on every write.",
            "in": "header",
            "name": "X-Agent-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The Figure 3 role performing this write. Required on every write.",
              "title": "X-Agent-Role"
            }
          },
          {
            "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Character"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvenanceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write a character's dossier",
        "tags": [
          "cast"
        ]
      }
    },
    "/cast/{id}/knowledge": {
      "get": {
        "description": "IF-03, `GET /cast/{id}/knowledge`. Every recorded state, each anchored to the scene it\nwas acquired in; deciding which of them a given scene may see is FR-OPS-01's job.",
        "operationId": "read_knowledge_cast__id__knowledge_get",
        "parameters": [
          {
            "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "maxLength": 120,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_-]*$",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeFile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read what a character knows",
        "tags": [
          "cast"
        ]
      },
      "put": {
        "description": "IF-04, `PUT /cast/{id}/knowledge` (canoniser).\n\nThe whole file, not a row: a character holds successive states on the same fact --\n`suspects` at 012, `knows` at 031 -- and an append-one-row route would invite the caller\nto drop the earlier state, which is the history FR-OPS-01 reads as-of.",
        "operationId": "write_knowledge_cast__id__knowledge_put",
        "parameters": [
          {
            "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Stable character id; it names the `cast/{id}/` directory. The grammar is FR-STORE-05's, declared here so a malformed id is a 422 at the edge rather than an exception from the store layer, and so the committed OpenAPI document (IF-08) states it.",
              "maxLength": 120,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_-]*$",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "description": "The Figure 3 role performing this write. Required on every write.",
            "in": "header",
            "name": "X-Agent-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The Figure 3 role performing this write. Required on every write.",
              "title": "X-Agent-Role"
            }
          },
          {
            "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`human` when a person is acting under the role, absent or `agent` otherwise. A human acts under a role, never beside it.",
              "title": "X-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvenanceRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write what a character knows",
        "tags": [
          "cast"
        ]
      }
    },
...
  }
}